+++ ospec/change-log.md
@@ -3,6 +3,7 @@
 
 ## Upcoming...
 _2018-xx-yy_
+- ospec: Test results now include `.message` and `.context` regardless of whether the test passed or failed. (#2227 @robertakarobin)
 <!-- Add new lines here. Version number will be decided later -->
 - Add `spy.calls` array property to get the `this` and `arguments` values for any arbitrary call.
 - Added `.throws` and `.notThrows` assertions to ospec. (#2255 @robertakarobin)